requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
python-dotenv
pytest
flask
requests
python-dotenv